+++ packages/manager/src/features/Linodes/LinodesDetail/LinodeConfigs/LinodeConfigDialog.tsx
@@ -44,7 +44,6 @@ import type { JSX } from 'react';
 
 import { FormLabel } from 'src/components/FormLabel';
 import { Link } from 'src/components/Link';
-import { DEFAULT_DEVICE_LIMIT } from 'src/constants';
 import { DeviceSelection } from 'src/features/Linodes/LinodesDetail/LinodeRescue/DeviceSelection';
 import { titlecase } from 'src/features/Linodes/presentation';
 import {
@@ -52,7 +51,6 @@ import {
   NATTED_PUBLIC_IP_HELPER_TEXT,
   NOT_NATTED_HELPER_TEXT,
 } from 'src/features/VPCs/constants';
-import { useFlags } from 'src/hooks/useFlags';
 import {
   handleFieldErrors,
   handleGeneralErrors,
@@ -70,7 +68,7 @@ import {
   StyledFormGroup,
   StyledRadioGroup,
 } from './LinodeConfigDialog.styles';
-import { getPrimaryInterfaceIndex } from './utilities';
+import { getPrimaryInterfaceIndex, useGetDeviceLimit } from './utilities';
 
 import type { ExtendedInterface } from '../LinodeSettings/InterfaceSelect';
 import type {
@@ -236,7 +234,6 @@ const deviceCounterDefault = 1;
 const finnixDiskID = 25669;
 
 export const LinodeConfigDialog = (props: Props) => {
-  const flags = useFlags();
   const formContainerRef = React.useRef<HTMLDivElement>(null);
   const { config, linodeId, onClose, open } = props;
 
@@ -246,9 +243,7 @@ export const LinodeConfigDialog = (props: Props) => {
     // eslint-disable-next-line no-console
     console.warn('Invalid memory value:', availableMemory);
   }
-  const deviceLimit = flags.blockStorageVolumeLimit
-    ? Math.max(DEFAULT_DEVICE_LIMIT, Math.min(availableMemory / 1024, 64))
-    : DEFAULT_DEVICE_LIMIT;
+  const deviceLimit = useGetDeviceLimit(availableMemory);
 
   const { isLinodeInterfacesEnabled } = useIsLinodeInterfacesEnabled();